--- a/src/lualib/ArrayConcat.ts
+++ b/src/lualib/ArrayConcat.ts
@@ -1,18 +1,22 @@
-function __TS__ArrayConcat(this: void, arr1: any[], ...args: any[]): any[] {
-    const out: any[] = [];
-    for (const val of arr1) {
-        out[out.length] = val;
+function __TS__ArrayConcat(this: void, arr: any[], items: any[]): any[] {
+    const result: any[] = [];
+    let len = 0;
+    for (const i of $range(1, arr.length)) {
+        len++;
+        result[len - 1] = arr[i - 1];
     }
-    for (const arg of args) {
-        if (Array.isArray(arg)) {
-            const argAsArray = arg;
-            for (const val of argAsArray) {
-                out[out.length] = val;
+    for (const i of $range(1, items.length)) {
+        const item = items[i - 1];
+        if (Array.isArray(item)) {
+            for (const j of $range(1, item.length)) {
+                len++;
+                result[len - 1] = item[j - 1];
             }
         } else {
-            out[out.length] = arg;
+            len++;
+            result[len - 1] = item;
         }
     }
 
-    return out;
+    return result;
 }