--- a/src/lualib/ArrayFilter.ts
+++ b/src/lualib/ArrayFilter.ts
@@ -4,9 +4,11 @@ function __TS__ArrayFilter<T>(
     callbackfn: (value: T, index?: number, array?: any[]) => boolean
 ): T[] {
     const result: T[] = [];
-    for (let i = 0; i < arr.length; i++) {
-        if (callbackfn(arr[i], i, arr)) {
-            result[result.length] = arr[i];
+    let len = 0;
+    for (const i of $range(1, arr.length)) {
+        if (callbackfn(arr[i - 1], i - 1, arr)) {
+            len++;
+            result[len - 1] = arr[i - 1];
         }
     }
     return result;