--- a/src/lualib/ArrayFind.ts
+++ b/src/lualib/ArrayFind.ts
@@ -2,11 +2,12 @@
 function __TS__ArrayFind<T>(
     this: void,
     arr: T[],
-    predicate: (value: T, index: number, obj: T[]) => unknown
+    predicate: (value: T, index: number, obj: T[]) => unknown,
+    thisArg?: any
 ): T | undefined {
     for (const i of $range(1, arr.length)) {
         const elem = arr[i - 1];
-        if (predicate(elem, i - 1, arr)) {
+        if (predicate.call(thisArg, elem, i - 1, arr)) {
             return elem;
         }
     }