--- a/src/lualib/ArrayFindIndex.ts
+++ b/src/lualib/ArrayFindIndex.ts
@@ -1,10 +1,11 @@
 function __TS__ArrayFindIndex<T>(
     this: void,
     arr: T[],
-    callbackFn: (element: T, index?: number, array?: T[]) => boolean
+    callbackFn: (element: T, index?: number, array?: T[]) => boolean,
+    thisArg?: any
 ): number {
     for (const i of $range(1, arr.length)) {
-        if (callbackFn(arr[i - 1], i - 1, arr)) {
+        if (callbackFn.call(thisArg, arr[i - 1], i - 1, arr)) {
             return i - 1;
         }
     }