--- a/src/lualib/ArrayFlat.ts
+++ b/src/lualib/ArrayFlat.ts
@@ -1,10 +1,23 @@
 function __TS__ArrayFlat(this: void, array: any[], depth = 1): any[] {
-    let result: any[] = [];
-    for (const value of array) {
+    const result: any[] = [];
+    let len = 0;
+    for (const i of $range(1, array.length)) {
+        const value = array[i - 1];
         if (depth > 0 && Array.isArray(value)) {
-            result = result.concat(__TS__ArrayFlat(value, depth - 1));
+            let toAdd: any[];
+            if (depth === 1) {
+                toAdd = value;
+            } else {
+                toAdd = value.flat(depth - 1);
+            }
+            for (const j of $range(1, toAdd.length)) {
+                const val = toAdd[j - 1];
+                len++;
+                result[len - 1] = val;
+            }
         } else {
-            result[result.length] = value;
+            len++;
+            result[len - 1] = value;
         }
     }