--- a/src/lualib/ArrayFlatMap.ts
+++ b/src/lualib/ArrayFlatMap.ts
@@ -1,12 +1,13 @@
 function __TS__ArrayFlatMap<T, U>(
     this: void,
     array: T[],
-    callback: (value: T, index: number, array: T[]) => U | readonly U[]
+    callback: (value: T, index: number, array: T[]) => U | readonly U[],
+    thisArg?: any
 ): U[] {
     const result: U[] = [];
     let len = 0;
     for (const i of $range(1, array.length)) {
-        const value = callback(array[i - 1], i - 1, array);
+        const value = callback.call(thisArg, array[i - 1], i - 1, array);
         if (Array.isArray(value)) {
             for (const j of $range(1, value.length)) {
                 len++;