--- a/src/lualib/ArrayFlatMap.ts
+++ b/src/lualib/ArrayFlatMap.ts
@@ -3,13 +3,18 @@ function __TS__ArrayFlatMap<T, U>(
     array: T[],
     callback: (value: T, index: number, array: T[]) => U | readonly U[]
 ): U[] {
-    let result: U[] = [];
-    for (let i = 0; i < array.length; i++) {
-        const value = callback(array[i], i, array);
-        if (type(value) === "table" && Array.isArray(value)) {
-            result = result.concat(value);
+    const result: U[] = [];
+    let len = 1;
+    for (const i of $range(1, array.length)) {
+        const value = callback(array[i - 1], i - 1, array);
+        if (Array.isArray(value)) {
+            for (const j of $range(1, value.length)) {
+                len++;
+                result[len - 1] = value[j - 1];
+            }
         } else {
-            result[result.length] = value as U;
+            len++;
+            result[len - 1] = value as U;
         }
     }