--- a/src/lualib/ArrayForEach.ts
+++ b/src/lualib/ArrayForEach.ts
@@ -3,7 +3,7 @@ function __TS__ArrayForEach<T>(
     arr: T[],
     callbackFn: (value: T, index?: number, array?: any[]) => any
 ): void {
-    for (let i = 0; i < arr.length; i++) {
-        callbackFn(arr[i], i, arr);
+    for (const i of $range(1, arr.length)) {
+        callbackFn(arr[i - 1], i - 1, arr);
     }
 }