--- a/src/lualib/ArrayIncludes.ts
+++ b/src/lualib/ArrayIncludes.ts
@@ -11,8 +11,8 @@ function __TS__ArrayIncludes<T>(this: T[], searchElement: T, fromIndex = 0): boo
         k = 0;
     }
 
-    for (const i of $range(k, len)) {
-        if (this[i] === searchElement) {
+    for (const i of $range(k + 1, len)) {
+        if (this[i - 1] === searchElement) {
             return true;
         }
     }