--- a/src/lualib/ArrayIndexOf.ts
+++ b/src/lualib/ArrayIndexOf.ts
@@ -1,31 +1,23 @@
-function __TS__ArrayIndexOf<T>(this: void, arr: T[], searchElement: T, fromIndex?: number): number {
+function __TS__ArrayIndexOf<T>(this: void, arr: T[], searchElement: T, fromIndex = 0): number {
     const len = arr.length;
     if (len === 0) {
         return -1;
     }
 
-    let n = 0;
-    if (fromIndex) {
-        n = fromIndex;
-    }
-
-    if (n >= len) {
+    if (fromIndex >= len) {
         return -1;
     }
 
-    let k: number;
-    if (n >= 0) {
-        k = n;
-    } else {
-        k = len + n;
-        if (k < 0) {
-            k = 0;
+    if (fromIndex < 0) {
+        fromIndex = len + fromIndex;
+        if (fromIndex < 0) {
+            fromIndex = 0;
         }
     }
 
-    for (let i = k; i < len; i++) {
-        if (arr[i] === searchElement) {
-            return i;
+    for (const i of $range(fromIndex + 1, len)) {
+        if (arr[i - 1] === searchElement) {
+            return i - 1;
         }
     }