--- a/src/lualib/ArrayJoin.ts
+++ b/src/lualib/ArrayJoin.ts
@@ -1,8 +1,7 @@
-function __TS__ArrayJoin(this: unknown[], separator = ",") {
-    let result = "";
-    for (const [index, value] of ipairs(this)) {
-        if (index > 1) result += separator;
-        result += value.toString();
+function __TS__ArrayJoin(this: void, arr: any[], separator = ",") {
+    const parts: string[] = [];
+    for (const i of $range(1, arr.length)) {
+        parts[i - 1] = arr[i - 1].toString();
     }
-    return result;
+    return table.concat(parts, separator);
 }