--- a/src/lualib/ArrayMap.ts
+++ b/src/lualib/ArrayMap.ts
@@ -1,7 +1,12 @@
-function __TS__ArrayMap<T, U>(this: void, arr: T[], callbackfn: (value: T, index?: number, array?: T[]) => U): U[] {
+function __TS__ArrayMap<T, U>(
+    this: void,
+    arr: T[],
+    callbackfn: (value: T, index?: number, array?: T[]) => U,
+    thisArg?: any
+): U[] {
     const result: U[] = [];
     for (const i of $range(1, arr.length)) {
-        result[i - 1] = callbackfn(arr[i - 1], i - 1, arr);
+        result[i - 1] = callbackfn.call(thisArg, arr[i - 1], i - 1, arr);
     }
     return result;
 }