--- a/src/lualib/ArrayMap.ts
+++ b/src/lualib/ArrayMap.ts
@@ -1,7 +1,7 @@
 function __TS__ArrayMap<T, U>(this: void, arr: T[], callbackfn: (value: T, index?: number, array?: T[]) => U): U[] {
-    const newArray: U[] = [];
-    for (let i = 0; i < arr.length; i++) {
-        newArray[i] = callbackfn(arr[i], i, arr);
+    const result: U[] = [];
+    for (const i of $range(1, arr.length)) {
+        result[i - 1] = callbackfn(arr[i - 1], i - 1, arr);
     }
-    return newArray;
+    return result;
 }