--- a/src/lualib/ArrayPush.ts
+++ b/src/lualib/ArrayPush.ts
@@ -1,6 +1,8 @@
-function __TS__ArrayPush<T>(this: void, arr: T[], ...items: T[]): number {
-    for (const item of items) {
-        arr[arr.length] = item;
+function __TS__ArrayPush<T>(this: void, arr: T[], items: T[]): number {
+    let len = arr.length;
+    for (const i of $range(1, items.length)) {
+        len++;
+        arr[len - 1] = items[i - 1];
     }
-    return arr.length;
+    return len;
 }