--- a/src/lualib/ArrayShift.ts
+++ b/src/lualib/ArrayShift.ts
@@ -1,3 +1,0 @@
-function __TS__ArrayShift<T>(this: void, arr: T[]): T {
-    return table.remove(arr, 1);
-}