--- a/src/lualib/ArraySome.ts
+++ b/src/lualib/ArraySome.ts
@@ -1,10 +1,11 @@
 function __TS__ArraySome<T>(
     this: void,
     arr: T[],
-    callbackfn: (value: T, index?: number, array?: any[]) => boolean
+    callbackfn: (value: T, index?: number, array?: any[]) => boolean,
+    thisArg?: any
 ): boolean {
     for (const i of $range(1, arr.length)) {
-        if (callbackfn(arr[i - 1], i - 1, arr)) {
+        if (callbackfn.call(thisArg, arr[i - 1], i - 1, arr)) {
             return true;
         }
     }