--- a/src/lualib/ArrayToObject.ts
+++ b/src/lualib/ArrayToObject.ts
@@ -1,7 +1,7 @@
 function __TS__ArrayToObject<T>(this: void, array: T[]): Record<number, T> {
     const object: Record<number, any> = {};
-    for (let i = 0; i < array.length; i += 1) {
-        object[i] = array[i];
+    for (const i of $range(1, array.length)) {
+        object[i - 1] = array[i - 1];
     }
     return object;
 }