--- a/src/lualib/ArrayUnshift.ts
+++ b/src/lualib/ArrayUnshift.ts
@@ -1,5 +1,6 @@
 function __TS__ArrayUnshift<T>(this: void, arr: T[], items: T[]): number {
     const length = items.length;
+    if (length === 0) return arr.length;
 
     for (const i of $range(arr.length, 1, -1)) {
         arr[i + length - 1] = arr[i - 1];