--- a/src/lualib/ObjectEntries.ts
+++ b/src/lualib/ObjectEntries.ts
@@ -1,7 +1,9 @@
 function __TS__ObjectEntries(this: void, obj: any): Array<string | number> {
     const result = [];
+    let len = 0;
     for (const key in obj) {
-        result[result.length] = [key, obj[key]];
+        len++;
+        result[len - 1] = [key, obj[key]];
     }
     return result;
 }