… benchmark/src/runtime_benchmarks/array_concat_array.ts
… benchmark/src/runtime_benchmarks/array_concat_spread.ts
… benchmark/src/runtime_benchmarks/array_every.ts
… benchmark/src/runtime_benchmarks/array_filter.ts
… benchmark/src/runtime_benchmarks/array_find.ts
… benchmark/src/runtime_benchmarks/array_findIndex.ts
… benchmark/src/runtime_benchmarks/array_flat.ts
… benchmark/src/runtime_benchmarks/array_flatMap.ts
… benchmark/src/runtime_benchmarks/array_foreach.ts
… benchmark/src/runtime_benchmarks/array_includes.ts
… benchmark/src/runtime_benchmarks/array_indexOf.ts
… benchmark/src/runtime_benchmarks/array_join.ts
… benchmark/src/runtime_benchmarks/array_map.ts
… benchmark/src/runtime_benchmarks/array_push_array.ts
… benchmark/src/runtime_benchmarks/array_push_multiple.ts
… benchmark/src/runtime_benchmarks/array_push_single.ts
… benchmark/src/runtime_benchmarks/array_reduce.ts
… benchmark/src/runtime_benchmarks/array_reduceRight.ts
… benchmark/src/runtime_benchmarks/array_reverse.ts
… benchmark/src/runtime_benchmarks/array_slice.ts
… benchmark/src/runtime_benchmarks/array_some.ts
… benchmark/src/runtime_benchmarks/array_splice.ts
… benchmark/src/runtime_benchmarks/array_unshift.ts
… src/lualib/ObjectAssign.ts
… src/lualib/ObjectKeys.ts
… src/lualib/ObjectValues.ts
… src/lualib/Spread.ts
… src/lualib/StringConcat.ts
… src/lualib/StringReplaceAll.ts
… src/lualib/StringSplit.ts