@GlassBricks 2021-09-08
The above commit addresses #1130; maybe could separate into different PR/branch